Add files via upload
This the code for the first steps of app. At first on clicking the app icon in mobile, a splash screen (poster image)will appear for 5 seconds then it movest to the page where different tasks are performed. The code for tasks are not yet ready.I will upload the apk file as soon as possible.